# backend/functions/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkling REST API
  version: 1.0.0
  description: >-
    Sparkling platform backend (SRS v1.0). All JSON, snake_case, ISO-8601 UTC timestamps, money in integer ZAR cents. Mutating routes accept `Idempotency-Key` (or `client_op_id` in the body); replays
    return the cached first response. Every response carries `X-Correlation-Id`. Clients send `X-Client-App` and `X-Client-Version`.
  contact:
    name: Sparkling engineering
    email: dev@sparkling.co.za
  license:
    name: Proprietary
    url: https://sparkling.co.za/legal
servers:
  - url: https://europe-west1-sparkling-4e89d.cloudfunctions.net/api/v1
    description: dev (Cloud Functions 2nd gen)
  - url: http://127.0.0.1:5001/sparkling-4e89d/europe-west1/api/v1
    description: Firebase emulator
...
security:
  - bearerAuth: []
paths:
...
  /quotations/{id}:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Quotation'
                  - type: object
                    properties:
                      work_order:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          ref:
                            type: string
                          status:
                            type: string
                        nullable: true
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Quotations
      operationId: getQuotation
      summary: Quotation with items, attachments (API-served urls), decision provenance, pdf_url and (staff) public_url
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
...
  /quotations/{id}/photos/{attachmentId}:
    get:
      responses:
        '200':
          description: Image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/heic:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Quotations
      operationId: getQuotationPhoto
      summary: 'Stream a damage photo (owner / outlet staff; Cache-Control: private, max-age=3600)'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
    delete:
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Quotations
      operationId: deleteQuotationPhoto
      summary: 'Staff: remove the object and the attachment row (before the decision; audited quotation.photo_delete)'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
...
  /quotations/{id}/pdf:
    get:
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Quotations
      operationId: getQuotationPdf
      summary: >-
        A4 PDF of the quotation (header, outlet, customer, vehicle, items, total, validity, status/decision, photo thumbnails, ref + QR of the public link); Content-Disposition: attachment;
        filename="<ref>.pdf"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
...
  /quotations/{id}/decision:
    post:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  quotation:
                    $ref: '#/components/schemas/Quotation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Quotations
      operationId: decideQuotation
      summary: >-
        Customer accepts or declines in the app (decision_source app). One-time: any later decision — in-app or via the public link — is 409 conflict {decided_at, status}; accepting after valid_until
        is 410 gone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  enum:
                    - accept
                    - decline
                note:
                  type: string
                  nullable: true
              required:
                - decision
...
  /public/quotations/{token}:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuotation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Public quote page
      operationId: getPublicQuotation
      summary: Public quotation view for the link holder (first name only, no phone / e-mail). 404 unknown token, 410 gone {ref,status} once the link expired
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Public link token from ${PUBLIC_WEB_BASE_URL}/q/<token>
  /public/quotations/{token}/photos/{attachmentId}:
    get:
      responses:
        '200':
          description: Image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/heic:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Public quote page
      operationId: getPublicQuotationPhoto
      summary: Stream a damage photo through the public link
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Public link token from ${PUBLIC_WEB_BASE_URL}/q/<token>
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /public/quotations/{token}/pdf:
    get:
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Public quote page
      operationId: getPublicQuotationPdf
      summary: Same PDF as GET /quotations/{id}/pdf, through the public link
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Public link token from ${PUBLIC_WEB_BASE_URL}/q/<token>
  /public/quotations/{token}/decision:
    post:
      responses:
        '200':
          description: Decided — returns the public view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuotation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Public quote page
      operationId: decidePublicQuotation
      summary: >-
        Accept or decline via the link (decision_source public_link, decision_by_name; audited quotation.decision_public with IP + user agent; quote_decided push to the assessor + outlet
        supervisors/managers, quote_decision_receipt to the customer). One-time: 409 conflict when already decided (in-app or via link), 410 gone when expired
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Public link token from ${PUBLIC_WEB_BASE_URL}/q/<token>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicDecision'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  responses:
    ValidationError:
      description: validation_error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: unauthenticated (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: forbidden (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: not_found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: conflict / invalid_transition (409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Gone:
      description: gone (410) — expired quote link or quotation past valid_until
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: rate_limited (429)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Internal:
      description: internal (500) — never includes stack traces
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - validation_error
                - conflict
                - invalid_transition
                - invalid_otp
                - gone
                - rate_limited
                - internal
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
          required:
            - code
            - message
            - correlation_id
      required:
        - error
...
    Quotation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ref:
          type: string
          example: QT-2026-0041
        customer_id:
          type: string
        vehicle_id:
          type: string
          format: uuid
        outlet_id:
          type: string
          format: uuid
        category:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - requested
            - assessing
            - quoted
            - accepted
            - declined
            - expired
            - converted
        amount_cents:
          type: integer
          nullable: true
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
          description: Alias of line_items
        items_note:
          type: string
          nullable: true
        assessor_id:
          type: string
          nullable: true
        valid_until:
          type: string
          format: date
          nullable: true
        quoted_at:
          type: string
          format: date-time
          nullable: true
        decided_at:
          type: string
          format: date-time
          nullable: true
        decision_by:
          type: string
          nullable: true
        decision_by_name:
          type: string
          nullable: true
        decision_note:
          type: string
          nullable: true
        decision_source:
          type: string
          enum:
            - app
            - public_link
            - staff
          nullable: true
        public_token_expires_at:
          type: string
          format: date-time
          nullable: true
        public_url:
          type: string
          description: 'Staff only: ${PUBLIC_WEB_BASE_URL}/q/<token>; null when no base URL is configured. The token itself is never listed.'
          nullable: true
        pdf_url:
          type: string
          example: /v1/quotations/<id>/pdf
        pdf_generated_at:
          type: string
          format: date-time
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/QuotationAttachment'
        client_op_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LineItem:
      type: object
      properties:
        label:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
          nullable: true
        category:
          type: string
          example: Bumper
          nullable: true
        service_id:
          type: string
          format: uuid
          nullable: true
        amount_cents:
          type: integer
          minimum: 0
          description: Line total (0 when a membership entitlement covers the service)
        quantity:
          type: integer
          minimum: 1
          nullable: true
        membership_benefit:
          type: string
          enum:
            - included
          nullable: true
        entitlement_id:
          type: string
          format: uuid
          nullable: true
      required:
        - label
        - amount_cents
...
    QuotationAttachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - damage_photo
            - document
        mime_type:
          type: string
          example: image/jpeg
        size_bytes:
          type: integer
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        caption:
          type: string
          nullable: true
        url:
          type: string
          description: 'API path that streams the file (signed-in: /v1/quotations/{id}/photos/{attachmentId}; public: /v1/public/quotations/{token}/photos/{attachmentId})'
        created_at:
          type: string
          format: date-time
      required:
        - id
        - kind
        - mime_type
        - size_bytes
        - url
...
    PublicQuotation:
      type: object
      properties:
        ref:
          type: string
        status:
          type: string
          enum:
            - requested
            - assessing
            - quoted
            - accepted
            - declined
            - expired
            - converted
        outlet:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
              nullable: true
            address_line:
              type: string
              nullable: true
          nullable: true
        customer:
          type: object
          properties:
            first_name:
              type: string
        vehicle:
          type: object
          properties:
            registration_no:
              type: string
            make:
              type: string
              nullable: true
            model:
              type: string
              nullable: true
            colour:
              type: string
              nullable: true
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        amount_cents:
          type: integer
          nullable: true
        currency:
          type: string
          example: ZAR
        valid_until:
          type: string
          format: date
          nullable: true
        quoted_at:
          type: string
          format: date-time
          nullable: true
        decided_at:
          type: string
          format: date-time
          nullable: true
        decision_source:
          type: string
          enum:
            - app
            - public_link
            - staff
          nullable: true
        expired:
          type: boolean
        can_decide:
          type: boolean
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              url:
                type: string
              caption:
                type: string
                nullable: true
              width:
                type: integer
                nullable: true
              height:
                type: integer
                nullable: true
        pdf_url:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - ref
        - status
        - customer
        - items
        - currency
        - expired
        - can_decide
        - attachments
        - pdf_url
    PublicDecision:
      type: object
      properties:
        decision:
          type: string
          enum:
            - accept
            - decline
        note:
          type: string
          maxLength: 500
          nullable: true
        accepted_by_name:
          type: string
          maxLength: 120
          nullable: true
      required:
        - decision
...